+++ apps/backoffice-v2/src/lib/blocks/variants/DefaultBlocks/hooks/useDefaultBlocksLogic/useDefaultBlocksLogic.tsx
@@ -282,6 +282,7 @@ export const useDefaultBlocksLogic = () => {
   const ubosRegistryProvidedBlock = useUbosRegistryProvidedBlock(
     ubosRegistryProvided,
     workflow?.context?.pluginsOutput?.ubo?.message,
+    workflow?.context?.pluginsOutput?.ubo?.isRequestTimedOut,
   );
 
   const directorsUserProvidedBlock = useDirectorsUserProvidedBlock(directorsUserProvided);