+++ packages/react-pdf-toolkit/src/templates/report/components/Section/components/Blocks/index.ts
@@ -1,3 +1,0 @@
-export * from './Block.Label';
-export * from './Blocks.Block';
-export * from './Blocks';