+++ packages/ui/src/components/organisms/DynamicForm/components/RSJVInputAdaters/TextInputAdapter/TextInputAdapter.tsx
@@ -6,9 +6,7 @@ import { FieldProps } from '@rjsf/utils';
 import { ComponentType, useMemo } from 'react';
 import { RJSFInputAdapter } from '@/components/organisms/DynamicForm/components/RSJVInputAdaters/types';
 
-export const TextInputAdapter:
-  | RJSFInputAdapter<string, unknown>
-  | RJSFInputAdapter<string | number, unknown> = (props: any) => {
+export const TextInputAdapter: RJSFInputAdapter<string, any> = props => {
   const InputComponent = useMemo(() => {
     const fieldsMap: Record<
       TextInputFieldType,