+++ services/workflows-service/src/common/types.ts
@@ -4,6 +4,7 @@ import z from 'zod';
 export type TDocumentWithoutPageType = Omit<DefaultContextSchema['documents'][number], 'pages'> & {
   pages: Array<Omit<DefaultContextSchema['documents'][number]['pages'][number], 'type'>>;
 };
+
 export type TDocumentsWithoutPageType = TDocumentWithoutPageType[];
 
 export const SubscriptionSchema = z.discriminatedUnion('type', [