+++ services/workflows-service/src/prisma/prisma.util.ts
@@ -1,8 +1,10 @@
-import { Prisma } from '@prisma/client';
-import { PrismaTransaction } from '@/types';
 import { PrismaService } from '@/prisma/prisma.service';
+import { PrismaTransaction } from '@/types';
+import { Prisma } from '@prisma/client';
 
 export const PRISMA_RECORD_NOT_FOUND_ERROR = 'P2025';
+export const PRISMA_FOREIGN_KEY_CONSTRAINT_ERROR = 'P2003';
+export const PRISMA_UNIQUE_CONSTRAINT_ERROR = 'P2002';
 
 export const isRecordNotFoundError = (
   error: unknown,
@@ -11,6 +13,16 @@ export const isRecordNotFoundError = (
 } =>
   error instanceof Prisma.PrismaClientKnownRequestError &&
   error.code === PRISMA_RECORD_NOT_FOUND_ERROR;
+
+export const isFkConstraintError = (
+  error: unknown,
+  fkName: string,
+): error is Prisma.PrismaClientKnownRequestError & {
+  code: typeof PRISMA_FOREIGN_KEY_CONSTRAINT_ERROR;
+} =>
+  error instanceof Prisma.PrismaClientKnownRequestError &&
+  error.code === PRISMA_FOREIGN_KEY_CONSTRAINT_ERROR &&
+  (error.meta as { field_name: string }).field_name.includes(fkName);
 
 export const transformStringFieldUpdateInput = async <
   T extends undefined | string | { set?: string },