+++ services/workflows-service/src/workflow/workflow.service.ts
@@ -1,5 +1,7 @@
 import { WorkflowTokenService } from '@/auth/workflow-token/workflow-token.service';
 import { BusinessRepository } from '@/business/business.repository';
+import { BusinessService } from '@/business/business.service';
+import { ajv } from '@/common/ajv/ajv.validator';
 import { AppLoggerService } from '@/common/app-logger/app-logger.service';
 import { EntityRepository } from '@/common/entity/entity.repository';
 import { SortOrder } from '@/common/query-filters/sort-order';
@@ -9,6 +11,13 @@ import { logDocumentWithoutId } from '@/common/utils/log-document-without-id/log
 import { CustomerService } from '@/customer/customer.service';
 import { EndUserRepository } from '@/end-user/end-user.repository';
 import { EndUserService } from '@/end-user/end-user.service';
+import { env } from '@/env';
+import { ValidationError } from '@/errors';
+import { PrismaService } from '@/prisma/prisma.service';
+import {
+  beginTransactionIfNotExistCurry,
+  defaultPrismaTransactionOptions,
+} from '@/prisma/prisma.util';
 import { ProjectScopeService } from '@/project/project-scope.service';
 import { FileService } from '@/providers/file/file.service';
 import { SalesforceService } from '@/salesforce/salesforce.service';
@@ -19,7 +28,9 @@ import type {
   TProjectId,
   TProjectIds,
 } from '@/types';
+import { UiDefinitionService } from '@/ui-definition/ui-definition.service';
 import { UserService } from '@/user/user.service';
+import { WorkflowDefinitionRepository } from '@/workflow-defintion/workflow-definition.repository';
 import { assignIdToDocuments } from '@/workflow/assign-id-to-documents';
 import { WorkflowAssigneeId } from '@/workflow/dtos/workflow-assignee-id';
 import { WorkflowDefinitionCloneDto } from '@/workflow/dtos/workflow-definition-clone';
@@ -34,6 +45,8 @@ import {
   DefaultContextSchema,
   getDocumentId,
   isErrorWithMessage,
+  isObject,
+  ProcessStatus,
 } from '@ballerine/common';
 import {
   ARRAY_MERGE_OPTION,
@@ -66,6 +79,7 @@ import {
 } from '@prisma/client';
 import { plainToClass } from 'class-transformer';
 import { isEqual, merge } from 'lodash';
+import mime from 'mime';
 import { WorkflowDefinitionCreateDto } from './dtos/workflow-definition-create';
 import { WorkflowDefinitionFindManyArgs } from './dtos/workflow-definition-find-many-args';
 import { WorkflowDefinitionUpdateInput } from './dtos/workflow-definition-update-input';
@@ -78,20 +92,8 @@ import {
   WorkflowRuntimeListQueryResult,
 } from './types';
 import { addPropertiesSchemaToDocument } from './utils/add-properties-schema-to-document';
-import { WorkflowDefinitionRepository } from '@/workflow-defintion/workflow-definition.repository';
 import { WorkflowEventEmitterService } from './workflow-event-emitter.service';
 import { WorkflowRuntimeDataRepository } from './workflow-runtime-data.repository';
-import mime from 'mime';
-import { env } from '@/env';
-import { ValidationError } from '@/errors';
-import { UiDefinitionService } from '@/ui-definition/ui-definition.service';
-import { ajv } from '@/common/ajv/ajv.validator';
-import { PrismaService } from '@/prisma/prisma.service';
-import {
-  beginTransactionIfNotExistCurry,
-  defaultPrismaTransactionOptions,
-} from '@/prisma/prisma.util';
-import { BusinessService } from '@/business/business.service';
 
 type TEntityId = string;
 
@@ -280,6 +282,31 @@ export class WorkflowService {
               workflow.workflowDefinition.documentsSchema,
             );
           },
+        ),
+        pluginsOutput: Object.fromEntries(
+          Object.entries(workflow.context.pluginsOutput || {}).map(([pluginName, pluginValue]) => {
+            if (
+              isObject(pluginValue) &&
+              pluginValue.status === ProcessStatus.IN_PROGRESS &&
+              pluginValue.invokedAt
+            ) {
+              const parsedDate = new Date(Number(pluginValue.invokedAt));
+
+              const TIMEOUT_IN_MS = 24 * 60 * 60 * 1000;
+
+              return [
+                pluginName,
+                {
+                  ...pluginValue,
+                  ...(parsedDate && !isNaN(parsedDate.getTime())
+                    ? { isRequestTimedOut: Date.now() - parsedDate.getTime() > TIMEOUT_IN_MS }
+                    : {}),
+                },
+              ];
+            }
+
+            return [pluginName, pluginValue];
+          }),
         ),
       },
       entity: getEntity(workflow),