+++ sdks/workflow-browser-sdk/CHANGELOG.md
@@ -1,5 +1,79 @@
 # @ballerine/workflow-browser-sdk
 
+## 0.6.2
+
+### Patch Changes
+
+- bumo
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.2
+  - @ballerine/common@0.9.1
+
+## 0.6.1
+
+### Patch Changes
+
+- ver bump
+- 9fe7a5c10: bump
+- Updated dependencies
+- Updated dependencies [9fe7a5c10]
+- Updated dependencies [9fe7a5c10]
+- Updated dependencies [9fe7a5c10]
+- Updated dependencies [9fe7a5c10]
+- Updated dependencies [9fe7a5c10]
+  - @ballerine/common@0.9.0
+  - @ballerine/workflow-core@0.6.1
+
+## 0.6.1-11210907c.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.0-11210907c.0
+  - @ballerine/workflow-core@0.6.1-11210907c.0
+
+## 0.6.0
+
+### Minor Changes
+
+- version bump
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.8.0
+  - @ballerine/workflow-core@0.6.0
+
+## 0.5.52
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.5.52
+
+## 0.5.51
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.5.51
+
+## 0.5.50
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.5.50
+
+## 0.5.49
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/workflow-core@0.5.49
+  - @ballerine/common@0.7.49
+
 ## 0.5.48
 
 ### Patch Changes